# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  gmod.ApiErrorResponse:
    properties:
      code:
        example: BAD_REQUEST
        type: string
      details: {}
      message:
        example: Invalid request
        type: string
      status:
        example: false
        type: boolean
    type: object
...
  gmod.BulkImportError:
    properties:
      field:
        type: string
      message:
        type: string
      recordKey:
        description: e.g. "AIBOX:channel:30"
        type: string
      row:
        type: integer
    type: object
  gmod.BulkImportResponse:
    properties:
      code:
        example: SUCCESS
        type: string
      errors:
        items:
          $ref: '#/definitions/gmod.BulkImportError'
        type: array
      ids:
        items:
          type: string
        type: array
      inserted:
        example: 5
        type: integer
      message:
        example: Bulk import completed
        type: string
      skipped:
        example: 1
        type: integer
      status:
        example: true
        type: boolean
      updated:
        example: 2
        type: integer
    type: object
...
info:
  contact: {}
  description: REST API for Gateway system
  title: Gateway API
  version: "1.0"
paths:
...
  /ingest/deviceManagement/export:
    get:
      description: Exports current device management records as XLSX file, scoped
        to active org.
      parameters:
      - description: Active Org ID
        in: header
        name: X-Active-Org
        required: true
        type: string
      - description: Filter by source family
        in: query
        name: sourceFamily
        type: string
      - description: Filter by entity type
        in: query
        name: entityType
        type: string
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
      security:
      - BearerAuth: []
      summary: Export device management data
      tags:
      - ingest-deviceManagement
  /ingest/deviceManagement/import:
    post:
      consumes:
      - multipart/form-data
      description: Parses an XLSX file and upserts device management records by business
        key.
      parameters:
      - description: Active Org ID
        in: header
        name: X-Active-Org
        required: true
        type: string
      - description: XLSX file
        in: formData
        name: file
        required: true
        type: file
      - default: "false"
        description: Preview mode
        in: formData
        name: dryRun
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gmod.BulkImportResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
      security:
      - BearerAuth: []
      summary: Bulk import device management from XLSX
      tags:
      - ingest-deviceManagement
  /ingest/deviceManagement/template:
    get:
      description: Returns an empty XLSX template prefilled with auth context for
        device management bulk import.
      parameters:
      - description: Active Org ID
        in: header
        name: X-Active-Org
        required: true
        type: string
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
      security:
      - BearerAuth: []
      summary: Download XLSX template
      tags:
      - ingest-deviceManagement
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"